+++ CHANGELOG.md
@@ -99,3 +99,4 @@ Bugfixing
 
 - Recognized Pages can also be TestPages -> Show a list where the user can select
 - User now has the possibility to use placeholders and then change the variable name to recommended var name
+- New shortcut Jump through Data items starting from buttom